........
  r29976 | res2002 | 2008-06-16 13:12:28 +0200 (m†, 16 jun 2008) | 1 line
  
  The Cg shader programs use stuff from the main plugin during cleanup, so keep a ref to it instead of a plain pointer
........
  r29978 | vknecht | 2008-06-16 18:31:01 +0200 (m†, 16 jun 2008) | 2 lines
  
  Eliminated a few compile warnings.
........
  r29985 | vknecht | 2008-06-16 21:45:00 +0200 (m†, 16 jun 2008) | 2 lines
  
  Added distfieldgen to svn:ignore property.
........
  r29990 | res2002 | 2008-06-17 15:10:20 +0200 (ti, 17 jun 2008) | 1 line
  
  RefTracker-related fixes
........
  r29991 | res2002 | 2008-06-17 15:11:04 +0200 (ti, 17 jun 2008) | 1 line
  
  Fix circular ref
........
  r29992 | res2002 | 2008-06-17 15:12:25 +0200 (ti, 17 jun 2008) | 1 line
  
  SV data allocation/deallocation fixes
........
  r29993 | res2002 | 2008-06-17 15:14:27 +0200 (ti, 17 jun 2008) | 1 line
  
  Make allocated memory blocks dump slightly more useful
........
  r29994 | res2002 | 2008-06-17 15:15:17 +0200 (ti, 17 jun 2008) | 1 line
  
  SV data allocation/deallocation fixes
........
  r29995 | res2002 | 2008-06-17 15:16:09 +0200 (ti, 17 jun 2008) | 1 line
  
  A bit more information
........
  r29996 | res2002 | 2008-06-17 16:49:15 +0200 (ti, 17 jun 2008) | 1 line
  
  Add missing ambient alias for lighting_default_binalpha; Fix alpha alias in lighting_default_shadowmap_binalpha
........


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/feature/newanim@29998 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/cssaver/cssaver.cpp
@@ -1592,7 +1592,7 @@ bool csSaver::SaveAllCollections()
 {
   // Get list of regions
   csRef<iCollectionArray> collections = engine->GetCollections ();
-  for (int i = 0; i < collections->GetSize (); i++)
+  for (size_t i = 0; i < collections->GetSize (); i++)
   {
     iCollection* collection = collections->Get (i);
     
@@ -1602,7 +1602,8 @@ bool csSaver::SaveAllCollections()
       scfInterfaceTraits<iSaverFile>::GetVersion ());
     csRef<iSaverFile> saverFile = scfQueryInterface<iSaverFile> (obj);
     
-    SaveCollectionFile (collection, saverFile->GetFile (), saverFile->GetFileType ());
+    SaveCollectionFile (collection, saverFile->GetFile (),
+      saverFile->GetFileType ());
   }
   
   return true;
